Just... a lot. So many changes.

+++ src/twig.c
@@ -26,7 +26,7 @@ EXPORT twig_dev_t *twig_open(void) {
     if (cedar->regs == MAP_FAILED)
         goto err_close;
 
-    if(twig_readl((uintptr_t)cedar->regs, VE_CTRL) & 0x00000001) {
+    if(twig_readl(cedar->regs, VE_CTRL) & 0x00000001) {
         fprintf(stderr, "WARNING: Cedar VE is still in H.264 mode, but twig_open was called again!\n");
         fprintf(stderr, "         Forcing a hardware reset in case the previous instance crashed!\n");
         ioctl(cedar->fd, IOCTL_SET_REFCOUNT, 0);
@@ -59,7 +59,7 @@ void *twig_get_ve_regs(twig_dev_t *cedar) {
         return NULL;
 
     if (cedar->active == 0) {
-        twig_writel((uintptr_t)cedar->regs, VE_CTRL, 0x00130001);
+        twig_writel(cedar->regs, VE_CTRL, 0x00130001);
         cedar->active = 1;
     }
     return cedar->regs;
@@ -80,7 +80,7 @@ void twig_put_ve_regs(twig_dev_t *cedar) {
     if (!cedar)
         return;
 
-    twig_writel((uintptr_t)cedar->regs, VE_CTRL, 0x00130007);
+    twig_writel(cedar->regs, VE_CTRL, 0x00130007);
     cedar->active = 0;
 }
